Move components around and add a prop to disable the viz selector
