+++ ui/packages/shared/profile/src/ProfileView/components/ActionButtons/SortByDropdown.tsx
@@ -18,7 +18,7 @@ import {
   FIELD_DIFF,
   FIELD_FUNCTION_NAME,
 } from '../../../ProfileIcicleGraph/IcicleGraphArrow';
-import {useProfileViewContext} from '../../../ProfileView/ProfileViewContext';
+import {useProfileViewContext} from '../../context/ProfileViewContext';
 
 const SortByDropdown = (): React.JSX.Element => {
   const [storeSortBy, setStoreSortBy] = useURLState('sort_by', {
