+++ ui/packages/shared/profile/src/ProfileView/components/TableColumnsDropdown.tsx
@@ -19,10 +19,10 @@ import {useURLState} from '@parca/components';
 import {ProfileType} from '@parca/parser';
 import {valueFormatter} from '@parca/utilities';
 
-import {useProfileViewContext} from '../../../ProfileView/ProfileViewContext';
-import {Row, isDummyRow} from '../../../Table';
-import ColumnsVisibility from '../../../Table/ColumnsVisibility';
-import {ColumnName, DataRow, addPlusSign, getRatioString} from '../../../Table/utils/functions';
+import {Row, isDummyRow} from '../../Table';
+import ColumnsVisibility from '../../Table/ColumnsVisibility';
+import {ColumnName, DataRow, addPlusSign, getRatioString} from '../../Table/utils/functions';
+import {useProfileViewContext} from '../context/ProfileViewContext';
 
 interface Props {
   profileType?: ProfileType;
