+++ ui/packages/shared/profile/src/ProfileView/components/VisualizationPanel.tsx
