+++ ui/packages/shared/profile/src/ProfileView/types/visualization.ts
@@ -71,4 +71,5 @@ export interface ProfileViewProps {
   compare?: boolean;
   onDownloadPProf: () => void;
   pprofDownloading?: boolean;
+  showVisualizationSelector?: boolean;
 }
