+++ ui/packages/shared/profile/src/components/VisualisationToolbar/MultiLevelDropdown.tsx
@@ -25,7 +25,7 @@ import {
   FIELD_DIFF,
   FIELD_FUNCTION_NAME,
 } from '../../ProfileIcicleGraph/IcicleGraphArrow';
-import {useProfileViewContext} from '../../ProfileView/ProfileViewContext';
+import {useProfileViewContext} from '../../ProfileView/context/ProfileViewContext';
 
 interface MenuItemType {
   label: string;
